# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/countries": {
            "post": {
                "tags": [
                    "Countries"
                ],
                "summary": "Insert countries into the database",
                "description": "Inserts countries into the database from an external API.",
                "operationId": "insertCountries",
                "responses": {
                    "200": {
                        "description": "Countries inserted successfully"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}